requirements.txt
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.9.0
azure-identity==1.14.0
azure-search-documents==11.4.0
azure-storage-blob==12.26.0
easyocr==1.7.2
fastapi==0.103.0
Flask==3.1.2
numpy==2.2.2
openai==1.75.0
pdf2image==1.17.0
pillow==10.4.0
pymupdf==1.25.3
pytesseract==0.3.13
python-dotenv==1.0.0
tiktoken==0.5.0
requests==2.32.3
torch==2.7.0
transformers==4.51.3
uvicorn==0.23.0